Lưu dữ liệu tạm thời vào file json.
